move TilingStrategy and TileInformation
