+++ operators/src/source/mod.rs
@@ -5,7 +5,5 @@ mod ogr_source;
 pub use self::csv::{
     CsvGeometrySpecification, CsvSource, CsvSourceParameters, CsvSourceStream, CsvTimeSpecification,
 };
-pub use self::gdal_source::{
-    GdalSource, GdalSourceParameters, GdalSourceProcessor, TilingStrategy,
-};
+pub use self::gdal_source::{GdalSource, GdalSourceParameters, GdalSourceProcessor};
 pub use self::ogr_source::{OgrSource, OgrSourceParameters};
